Show a loading spinner until chats have been fetched
