+++ src/features/Chat/ChatMenu.tsx
@@ -1,10 +1,8 @@
 import styled, { css } from 'styled-components';
-import { useEffect, useState } from 'react';
+import { useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useDispatch, useSelector } from 'react-redux';
 import {
-  ChatContact,
-  addChat,
   getChatsByActiveCategory,
   setActiveCategory,
   setActiveChat,
@@ -50,57 +48,6 @@ const ChatMenu = () => {
     }
     return '';
   };
-
-  // Fetch chats from API
-  useEffect(() => {
-    // This function calls the API to get a list of chats (contacts and messages)
-    // The function is called when the user first logs in, and whenever they would like to refresh their chat list
-    // It gets a list of contacts for the user, and then gets a list of messages between the user and each contact
-    // It then combines the contact and message lists to create a list of chats
-    const fetchChats = async () => {
-      const maxMessagesAtOnce = 20;
-
-      const response = await fetch('/api/myuser');
-      if (response.ok) {
-        const user = await response.json();
-        const userId = user['user']['id'];
-        const contactData = await fetch(`api/users/${userId}/contacts`);
-        const contactsDataJson = await contactData.json();
-        const contacts = contactsDataJson['resources'];
-        const contactIds = contacts.map(c => c.id).join(',');
-
-        const messagesData = await fetch(
-          `api/users/${userId}/messages?contact_user_ids=${contactIds}&max=${maxMessagesAtOnce}&desc=true`,
-        );
-        const messagesDataJson = await messagesData.json();
-        const messages = messagesDataJson['resources'];
-
-        const chats: ChatContact[] = contacts.map(contact => {
-          const messageList = messages.filter(
-            message =>
-              message['recipient_id'] === contact.id ||
-              message['sender_id'] === contact.id,
-          );
-          return {
-            active: contact.active,
-            category: 'active',
-            displayName: contact.display_name,
-            id: contact.id,
-            messages: messageList,
-            name: contact.name,
-            role: contact.role,
-            status: contact.status,
-          };
-        });
-        chats.forEach((chat: ChatContact) => dispatch(addChat(chat)));
-        // Set the active chat to the first chat in the list
-        if (chats.length) {
-          dispatch(setActiveChat(chats[0].id));
-        }
-      }
-    };
-    fetchChats();
-  }, []);
 
   // The chooseChat function is called when the user clicks on a chat in the chat menu
   // It sets the active chat to the chat that was clicked on
